requirements.txt
PyQt5>=5.15.4
requests
fastapi>=0.104.1
uvicorn>=0.24.0
beautifulsoup4>=4.12.2
charset-normalizer>=3.3.2
python-dotenv>=1.0.0